+++ app/src/main/java/com/example/inventarisbarang/dao/RuanganDao.kt
@@ -23,5 +23,5 @@ interface RuanganDao {
     fun getAllRuangan(): LiveData<List<Ruangan>>
 
     @Query("SELECT * FROM ruangan WHERE id = :ruanganId")
-    fun getRuanganById(ruanganId: Int): LiveData<Ruangan>
+    fun getRuanganById(ruanganId: Long): LiveData<Ruangan>
 }
