+++ app/src/main/java/com/example/inventarisbarang/entity/Barang.kt
@@ -5,12 +5,12 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "barang")
 data class Barang(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val nama: String,
     val kategori: String,
     val jumlah: Int,
     val tanggalMasuk: String,
     val kondisi: String,
-    val ruanganId: Int,
-    val karyawanId: Int
+    val ruanganId: Long,
+    val karyawanId: Long
 )
